% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalConditionalLik_wn2.R
\name{evalConditionalLik_wn2}
\alias{evalConditionalLik_wn2}
\title{Internal use only}
\usage{
evalConditionalLik_wn2(
  Z,
  S1,
  R1,
  U,
  V,
  Y,
  Eta,
  W_adj,
  Lambda_a,
  phi,
  EE,
  log_t_k,
  t_0,
  alpha,
  alpha_0 = 1.1,
  beta_0 = 25
)
}
\description{
Internal use only
}
