% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICL.R
\name{ICL}
\alias{ICL}
\title{Integrated Complete Likelihood (ICL) for wecan objects}
\usage{
ICL(object)
}
\arguments{
\item{object}{Object of class wecan}
}
\description{
Integrated Complete Likelihood (ICL) for wecan objects
}
