% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wecan.R
\name{wecan}
\alias{wecan}
\title{wecan: Weighted Edge Clustering Adjusting for Noise}
\usage{
wecan(
  A,
  K = 2,
  p = 3,
  maxIter = 1000,
  maxIterVB = 100,
  a_0 = 1,
  A_0 = 3,
  B_0 = 0.5,
  nu_0 = 2,
  eta_0 = 5,
  CGSteps = 25,
  distribution = "Normal",
  weight_attr_name = "weight",
  eps = 1e-05,
  upsilon_SR1 = 2,
  upsilon_SR2 = 2,
  upsilon_UV = 2,
  Phi_SR1 = toeplitz((2:1)/2),
  Phi_SR2 = toeplitz((2:1)/2),
  Phi_UV = toeplitz((2:1)/2),
  noise_mean = 0.1,
  alpha_init = NULL,
  a_a = 1,
  b_a = 200,
  t_0 = 0.05,
  alpha_0 = 1.1,
  beta_0 = 25
)
}
\arguments{
\item{A}{Network object of class igraph}

\item{K}{integer. Maximum number of clusters}

\item{p}{integer. Dimension of latent space.}

\item{maxIter}{integer. Maximum number of iterations of the VB-EM algorithm}

\item{maxIterVB}{integer. Maximum number of iterations of the
VB step nested within the EM algorithm}

\item{a_0}{description}

\item{A_0}{description}

\item{B_0}{description}

\item{nu_0}{description}

\item{eta_0}{description}

\item{CGSteps}{Number of steps in each conjugate gradient update
nested within the VB-EM algorithm}

\item{distribution}{character.  Currently, only "Normal" and "Poisson" are supported.}

\item{weight_attr_name}{character.  Name of the weight attribute of A to be modeled.}

\item{eps}{description}

\item{upsilon_SR1}{description}

\item{upsilon_SR2}{description}

\item{upsilon_UV}{description}

\item{Phi_SR1}{description}

\item{Phi_SR2}{description}

\item{Phi_UV}{description}

\item{a_a}{description}

\item{b_a}{description}

\item{t_0}{description}

\item{alpha_0}{description}

\item{beta_0}{description}
}
\value{
Object of class 'wecan' with the following elements:
\itemize{
\item estimates, a list with the following elements:
\itemize{
\item S1 description
\item R1 description
\item S2 description
\item R2 description
\item U description
\item V description
\item Y description
\item Eta description
\item Pmk description
\item alpha description
\item Elog_t_k description
\item E_t_0 description
\item phi description
\item sigma_SR1 description
\item sigma_SR2 description
\item sigma_UV description
\item Lambda description
\item Z_est description
\item Z_init description
}
\item userInputs, a list with the following elements:
\itemize{
\item K description
\item p description
\item Phi_SR1 description
\item Phi_SR2 description
\item Phi_UV description
\item upsilon_SR1 description
\item upsilon_SR2 description
\item upsilon_UV description
\item a_0 description
\item A_0 description
\item B_0 description
\item nu_0 description
\item eta_0 description
\item lambda_a description
}
\item network, the original igraph object
}
}
\description{
This function implements the weighted edge clustering algorithm
of Li and Sewell
}
\examples{
\dontrun{
#Fill in some code here to show how to use wecan(), BIC(), and ICL().
}

}
