% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wecan_sim.R
\name{wecan_sim}
\alias{wecan_sim}
\title{Title: Create the simulated networks
Description: This function created simulated networks with clustered edges.}
\usage{
wecan_sim(
  p = 2,
  K = 2 * p,
  n = 400,
  Lambda_sc = 5,
  M = round(n * (n - 1) * 0.05),
  conc = c(1.5, 2.5, 6, 25, 150, 200, 1.75, 5, 10, 50, 150, 200)[6 * (p - 2) + K - 1],
  UVShape = 20,
  UVRate = 4,
  if_weight = TRUE,
  mag = 1,
  Ymag = 1,
  with_noise = TRUE,
  noise_mean = 0.5,
  noise_prop = 0.1,
  distribution = "Normal",
  phi_min = 0.05,
  phi_max = 0.5,
  beta = c(4, 0, 2, 13)
)
}
\arguments{
\item{p}{integer. Number of latent dimensions}

\item{K}{integer. True number of clusters}

\item{n}{integer. Number of actors in the network}

\item{M}{integer. Number of edges in the network}

\item{conc}{concentration parameter of the mixtures of von Mises-Fisher distributions to generate U, V}

\item{UVShape}{shape parameter of the Gamma distribution to generate the magnitude of U, V}

\item{UVRate}{rate parameter of the Gamma distribution to generate the magnitude of U, V}

\item{if_weight}{logical. indicates whether simulate a weighted network}

\item{with_noise}{logical. indicates whether simulate noisy edges}

\item{noise_mean}{numeric. mean weights for the noisy edges}

\item{distribution}{character. The distribution for network's weights, currently only "Normal" can be used.}

\item{beta}{vector. mean weights for edges in each cluster}
}
\value{
simulated networks, edge clusters and the nodes-centric/edge-centric features
}
\description{
Title: Create the simulated networks
Description: This function created simulated networks with clustered edges.
}
